refactor!: drop MaybeDisjoint::Consumed; move sets into sets/ module
Remove the `Consumed` variant from `MaybeDisjoint`. The empty set is now
canonically `Connected(EnumInterval::empty())`, which doubles as the
drained-iterator sentinel. Eliminates the dual identity where `Consumed`
and `Connected(empty)` denoted the same set but differed at the type
level.

Split `sets.rs` into `sets/{finite,half,enum_interval,disjoint}.rs` and
absorb the old top-level `disjoint` module. `MaybeDisjoint` now lives at
`crate::sets::MaybeDisjoint`.
